# [libc++] Optimize ranges::swap_ranges for vector<bool>::iterator