This PR optimizes the performance of `std::ranges::swap_ranges` for `vector<bool>::iterator`, addressing a subtask outlined in issue #64038. The optimizations yield performance improvements of up to **611x** for aligned range swap and **78x** for unaligned range swap comparison. Additionally, comprehensive tests covering up to 4 storage words (256 bytes) with odd and even bit sizes are provided, which validate the proposed optimizations in this patch.

```
-----------------------------------------------------------------------------------
Benchmark                                         Before        After   Improvement
-----------------------------------------------------------------------------------
bm_ranges_swap_ranges_vb_aligned/8               18.7 ns      4.19 ns           4x
bm_ranges_swap_ranges_vb_aligned/16              34.9 ns      4.30 ns           8x
bm_ranges_swap_ranges_vb_aligned/32              67.0 ns      4.20 ns          16x
bm_ranges_swap_ranges_vb_aligned/64               129 ns      3.93 ns          33x
bm_ranges_swap_ranges_vb_aligned/128              287 ns      4.34 ns          66x
bm_ranges_swap_ranges_vb_aligned/256              590 ns      5.41 ns         109x
bm_ranges_swap_ranges_vb_aligned/512             1173 ns      7.13 ns         165x
bm_ranges_swap_ranges_vb_aligned/1024            2358 ns      6.76 ns         349x
bm_ranges_swap_ranges_vb_aligned/2048            4737 ns      9.93 ns         492x
bm_ranges_swap_ranges_vb_aligned/4096            9206 ns      16.7 ns         551x
bm_ranges_swap_ranges_vb_aligned/8192           17524 ns      31.1 ns         563x
bm_ranges_swap_ranges_vb_aligned/16384          34151 ns      55.9 ns         611x
bm_ranges_swap_ranges_vb_aligned/32768          66369 ns       115 ns         577x
bm_ranges_swap_ranges_vb_aligned/65536         133778 ns       240 ns         557x
bm_ranges_swap_ranges_vb_aligned/131072        285673 ns       480 ns         595x
bm_ranges_swap_ranges_vb_aligned/262144        613082 ns      1088 ns         563x
bm_ranges_swap_ranges_vb_aligned/524288       1237831 ns      2186 ns         566x
bm_ranges_swap_ranges_vb_aligned/1048576      2426294 ns      4334 ns         560x
bm_ranges_swap_ranges_vb_unaligned/8             17.0 ns      10.0 ns         1.7x
bm_ranges_swap_ranges_vb_unaligned/64             149 ns      12.7 ns          12x
bm_ranges_swap_ranges_vb_unaligned/512           1131 ns      23.8 ns          48x
bm_ranges_swap_ranges_vb_unaligned/4096          8801 ns       128 ns          69x
bm_ranges_swap_ranges_vb_unaligned/32768        71029 ns       943 ns          75x
bm_ranges_swap_ranges_vb_unaligned/262144      563353 ns      8068 ns          70x
bm_ranges_swap_ranges_vb_unaligned/1048576    2497529 ns     32009 ns          78x
```